+++ tests/Regressors/GradientBoostTest.php
@@ -22,7 +22,7 @@ class GradientBoostTest extends TestCase
 {
     protected const TRAIN_SIZE = 400;
     protected const TEST_SIZE = 10;
-    protected const MIN_SCORE = 0.7;
+    protected const MIN_SCORE = 0.9;
 
     protected const RANDOM_SEED = 0;