+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Added Random, K-MC2, and Plus Plus seeders
     - Accelerated Mean Shift with Ball Tree
     - Added radius estimation to Mean Shift
+    - K Means and Mean Shift now implement Probabilistic
     - Gaussian Mixture now supports seeders
     - Changed order of K Means hyperparameters
     - Moved Ranking interface to anomaly detector namespace